refactor: split memory transcript corpus module
