+++ packages/memory-host-sdk/src/host/session-files.ts
@@ -6,10 +6,8 @@ import { readRegularFile, statRegularFile } from "./fs-utils.js";
 import { hashText } from "./hash.js";
 import { createSubsystemLogger, redactSensitiveText } from "./openclaw-runtime-io.js";
 import {
-  getRuntimeConfig,
   HEARTBEAT_PROMPT,
   HEARTBEAT_TOKEN,
-  canonicalizeMainSessionAlias,
   hasInterSessionUserProvenance,
   isCompactionCheckpointTranscriptFileName,
   isCronRunSessionKey,
@@ -18,18 +16,24 @@ import {
   isSessionArchiveArtifactName,
   isSilentReplyPayloadText,
   isUsageCountedSessionTranscriptFileName,
-  listSessionEntries,
   parseUsageCountedSessionIdFromFileName,
-  resolveSessionFilePath,
-  resolveStorePath,
-  resolveSessionAgentId,
   resolveSessionTranscriptsDirForAgent,
   stripInboundMetadata,
   stripInternalRuntimeContext,
-  type SessionEntry,
 } from "./openclaw-runtime-session.js";
 import { retryTransientMemoryRead } from "./read-retry.js";
+import {
+  listSessionTranscriptCorpusEntriesForAgent,
+  listSessionTranscriptCorpusEntriesForAgentSync,
+  type SessionTranscriptCorpusEntry,
+} from "./session-transcript-corpus.js";
 import type { MemorySessionSyncTarget } from "./types.js";
+
+export {
+  listSessionTranscriptCorpusEntriesForAgent,
+  type SessionTranscriptCorpusArtifactKind,
+  type SessionTranscriptCorpusEntry,
+} from "./session-transcript-corpus.js";
 
 const DREAMING_NARRATIVE_RUN_PREFIX = "dreaming-narrative-";
 // Keep the historical one-line-per-message export shape for normal turns, but
@@ -72,23 +76,6 @@ export type SessionTranscriptClassification = {
   cronRunTranscriptPaths: ReadonlySet<string>;
 };
 
-export type SessionTranscriptCorpusArtifactKind =
-  | "active-session"
-  | "archive-artifact"
-  | "orphan-file-artifact";
-
-export type SessionTranscriptCorpusEntry = {
-  agentId: string;
-  sessionFile: string;
-  sessionId: string;
-  artifactKind: SessionTranscriptCorpusArtifactKind;
-  sessionKey?: string;
-  /** True when this transcript belongs to an internal dreaming narrative run. */
-  generatedByDreamingNarrative?: boolean;
-  /** True when this transcript belongs to an isolated cron run session. */
-  generatedByCronRun?: boolean;
-};
-
 export type ResolvedMemorySessionSyncTarget = {
   agentId: string;
   sessionFile: string;
@@ -104,11 +91,6 @@ export type ResolvedSessionTranscriptIdentity = {
 type SessionTranscriptStoreEntry = {
   sessionFile?: unknown;
   sessionId?: unknown;
-};
-
-type SessionEntrySummary = {
-  sessionKey: string;
-  entry: SessionEntry;
 };
 
 function shouldSkipTranscriptFileForDreaming(absPath: string): boolean {
@@ -209,10 +191,6 @@ function isDreamingNarrativeSessionStoreKey(sessionKey: string): boolean {
   return sessionSegment.startsWith(DREAMING_NARRATIVE_RUN_PREFIX);
 }
 
-function isDreamingNarrativeSessionKeyLike(value: unknown): boolean {
-  return typeof value === "string" && isDreamingNarrativeSessionStoreKey(value);
-}
-
 function hasCronRunSessionKey(value: unknown): boolean {
   return typeof value === "string" && isCronRunSessionKey(value);
 }
@@ -242,26 +220,8 @@ function normalizeComparablePath(pathname: string): string {
   return process.platform === "win32" ? resolved.toLowerCase() : resolved;
 }
 
-function normalizeRealComparablePath(pathname: string): string {
-  try {
-    return normalizeComparablePath(fsSync.realpathSync(pathname));
-  } catch {
-    try {
-      return normalizeComparablePath(
-        path.join(fsSync.realpathSync(path.dirname(pathname)), path.basename(pathname)),
-      );
-    } catch {
-      return normalizeComparablePath(pathname);
-    }
-  }
-}
-
 export function normalizeSessionTranscriptPathForComparison(pathname: string): string {
   return normalizeComparablePath(pathname);
-}
-
-function rememberArtifactDir(dirs: Map<string, string>, dir: string): void {
-  dirs.set(normalizeRealComparablePath(dir), dir);
 }
 
 function resolveSessionStoreTranscriptPath(
@@ -286,74 +246,6 @@ function resolveSessionStoreTranscriptResolvedPath(
   return null;
 }
 
-function resolveSessionStoreTranscriptCorpusPath(
-  agentId: string,
-  sessionsDir: string,
-  entry: { sessionFile?: unknown; sessionId?: unknown } | undefined,
-): string | null {
-  const sessionFile =
-    typeof entry?.sessionFile === "string" && entry.sessionFile.trim().length > 0
-      ? entry.sessionFile.trim()
-      : undefined;
-  const sessionId =
-    typeof entry?.sessionId === "string" && entry.sessionId.trim().length > 0
-      ? entry.sessionId.trim()
-      : sessionFile
-        ? parseUsageCountedSessionIdFromFileName(path.basename(sessionFile))
-        : null;
-  if (!sessionId) {
-    return null;
-  }
-  try {
-    if (!sessionFile) {
-      return resolveSessionFilePath(sessionId, undefined, { agentId, sessionsDir });
-    }
-    const resolved = resolveSessionFilePath(
-      sessionId,
-      { sessionFile },
-      {
-        agentId,
-        sessionsDir,
-      },
-    );
-    if (!path.isAbsolute(sessionFile)) {
-      const candidate = path.resolve(sessionsDir, sessionFile);
-      if (
-        normalizeComparablePath(path.dirname(candidate)) !== normalizeComparablePath(sessionsDir)
-      ) {
-        return null;
-      }
-      return normalizeRealComparablePath(resolved) === normalizeRealComparablePath(candidate)
-        ? candidate
-        : null;
-    }
-    const pathAgentId = extractAgentIdFromSessionPath(sessionFile);
-    if (pathAgentId && normalizeAgentId(pathAgentId) !== normalizeAgentId(agentId)) {
-      return null;
-    }
-    return normalizeRealComparablePath(resolved) === normalizeRealComparablePath(sessionFile)
-      ? sessionFile
-      : null;
-  } catch {
-    return null;
-  }
-}
-
-function classifySessionEntry(
-  sessionKey: string,
-  entry: SessionEntry,
-): {
-  generatedByDreamingNarrative: boolean;
-  generatedByCronRun: boolean;
-} {
-  return {
-    generatedByDreamingNarrative:
-      isDreamingNarrativeSessionStoreKey(sessionKey) ||
-      isDreamingNarrativeSessionKeyLike(entry.spawnedBy),
-    generatedByCronRun: isCronRunSessionKey(sessionKey) || hasCronRunSessionKey(entry.spawnedBy),
-  };
-}
-
 function extractAgentIdFromSessionsDir(sessionsDir: string): string | null {
   const parts = path.normalize(path.resolve(sessionsDir)).split(path.sep).filter(Boolean);
   const sessionsIndex = parts.length - 1;
@@ -372,214 +264,6 @@ function isCanonicalSessionsDirForAgent(sessionsDir: string, agentId: string): b
     normalizeComparablePath(sessionsDir) ===
     normalizeComparablePath(resolveSessionTranscriptsDirForAgent(agentId))
   );
-}
-
-function isRegularSessionTranscriptFile(absPath: string): boolean {
-  try {
-    return fsSync.lstatSync(absPath).isFile();
-  } catch {
-    return false;
-  }
-}
-
-function toSessionStoreCorpusEntry(
-  agentId: string,
-  sessionsDir: string,
-  summary: SessionEntrySummary,
-): SessionTranscriptCorpusEntry | null {
-  const sessionFile = resolveSessionStoreTranscriptCorpusPath(agentId, sessionsDir, summary.entry);
-  if (!sessionFile || !isUsageCountedSessionTranscriptFileName(path.basename(sessionFile))) {
-    return null;
-  }
-  const sessionId =
-    typeof summary.entry.sessionId === "string" && summary.entry.sessionId.trim()
-      ? summary.entry.sessionId.trim()
-      : parseUsageCountedSessionIdFromFileName(path.basename(sessionFile));
-  if (!sessionId) {
-    return null;
-  }
-  const sessionKey = summary.sessionKey.trim();
-  const classification = classifySessionEntry(summary.sessionKey, summary.entry);
-  return {
-    agentId,
-    artifactKind: "active-session",
-    sessionFile,
-    sessionId,
-    ...(sessionKey ? { sessionKey } : {}),
-    ...(classification.generatedByDreamingNarrative ? { generatedByDreamingNarrative: true } : {}),
-    ...(classification.generatedByCronRun ? { generatedByCronRun: true } : {}),
-  };
-}
-
-function listSessionTranscriptArtifactFiles(sessionsDir: string): string[] {
-  try {
-    return fsSync
-      .readdirSync(sessionsDir, { withFileTypes: true })
-      .filter((entry) => entry.isFile())
-      .map((entry) => entry.name)
-      .filter((name) => isUsageCountedSessionTranscriptFileName(name))
-      .map((name) => path.join(sessionsDir, name));
-  } catch {
-    return [];
-  }
-}
-
-function classifyTranscriptArtifact(
-  artifactPath: string,
-  activeEntriesByPath: ReadonlyMap<string, SessionTranscriptCorpusEntry>,
-): {
-  generatedByDreamingNarrative: boolean;
-  generatedByCronRun: boolean;
-} {
-  const directEntry = activeEntriesByPath.get(normalizeRealComparablePath(artifactPath));
-  if (directEntry) {
-    return {
-      generatedByDreamingNarrative: directEntry.generatedByDreamingNarrative === true,
-      generatedByCronRun: directEntry.generatedByCronRun === true,
-    };
-  }
-  const sessionsDir = path.dirname(artifactPath);
-  const primarySessionId = parseUsageCountedSessionIdFromFileName(path.basename(artifactPath));
-  const primaryEntry =
-    primarySessionId && isSessionArchiveArtifactName(path.basename(artifactPath))
-      ? activeEntriesByPath.get(
-          normalizeRealComparablePath(path.join(sessionsDir, `${primarySessionId}.jsonl`)),
-        )
-      : undefined;
-  return {
-    generatedByDreamingNarrative: primaryEntry?.generatedByDreamingNarrative === true,
-    generatedByCronRun: primaryEntry?.generatedByCronRun === true,
-  };
-}
-
-function toArtifactCorpusEntry(
-  agentId: string,
-  artifactPath: string,
-  activeEntriesByPath: ReadonlyMap<string, SessionTranscriptCorpusEntry>,
-): SessionTranscriptCorpusEntry | null {
-  const sessionId = parseUsageCountedSessionIdFromFileName(path.basename(artifactPath));
-  if (!sessionId) {
-    return null;
-  }
-  const artifactKind = isSessionArchiveArtifactName(path.basename(artifactPath))
-    ? "archive-artifact"
-    : "orphan-file-artifact";
-  const classification = classifyTranscriptArtifact(artifactPath, activeEntriesByPath);
-  return {
-    agentId,
-    artifactKind,
-    sessionFile: artifactPath,
-    sessionId,
-    ...(classification.generatedByDreamingNarrative ? { generatedByDreamingNarrative: true } : {}),
-    ...(classification.generatedByCronRun ? { generatedByCronRun: true } : {}),
-  };
-}
-
-function listSessionTranscriptCorpusEntriesForAgentSync(
-  agentId: string,
-): SessionTranscriptCorpusEntry[] {
-  const normalizedAgentId = normalizeAgentId(agentId);
-  const cfg = getRuntimeConfig();
-  const configuredStore = cfg.session?.store;
-  const storePath = resolveStorePath(configuredStore, {
-    agentId: normalizedAgentId,
-  });
-  const sessionsDir = path.dirname(storePath);
-  const fixedStoreOwnerAgentId = extractAgentIdFromSessionsDir(sessionsDir);
-  const isAgentOwnedFixedStore =
-    fixedStoreOwnerAgentId !== null &&
-    normalizeAgentId(fixedStoreOwnerAgentId) === normalizedAgentId;
-  const isSharedFixedStore =
-    typeof configuredStore === "string" &&
-    configuredStore.trim().length > 0 &&
-    !configuredStore.includes("{agentId}") &&
-    !isAgentOwnedFixedStore;
-  const activeEntriesByPath = new Map<string, SessionTranscriptCorpusEntry>();
-  const activeEntryOwnersByPath = new Map<string, string>();
-  const artifactDirsByPath = new Map<string, string>();
-  rememberArtifactDir(artifactDirsByPath, sessionsDir);
-  for (const summary of listSessionEntries({
-    agentId: normalizedAgentId,
-    hydrateSkillPromptRefs: false,
-    storePath,
-  })) {
-    const sessionKey = isSharedFixedStore
-      ? summary.sessionKey
-      : canonicalizeMainSessionAlias({
-          cfg,
-          agentId: normalizedAgentId,
-          sessionKey: summary.sessionKey,
-        });
-    const ownerAgentId = resolveSessionAgentId({
-      config: cfg,
-      sessionKey,
-      ...(isSharedFixedStore ? {} : { fallbackAgentId: normalizedAgentId }),
-    });
-    const entry = toSessionStoreCorpusEntry(ownerAgentId, sessionsDir, summary);
-    if (!entry) {
-      continue;
-    }
-    const normalizedEntryPath = normalizeRealComparablePath(entry.sessionFile);
-    activeEntryOwnersByPath.set(normalizedEntryPath, ownerAgentId);
-    rememberArtifactDir(artifactDirsByPath, path.dirname(entry.sessionFile));
-    if (ownerAgentId === normalizedAgentId) {
-      activeEntriesByPath.set(normalizedEntryPath, entry);
-    }
-  }
-  const includeUnownedArtifacts = !isSharedFixedStore;
-  const corpusEntries = [...activeEntriesByPath.values()].filter((entry) =>
-    isRegularSessionTranscriptFile(entry.sessionFile),
-  );
-  const scannedArtifactPaths = new Set<string>();
-  for (const artifactDir of artifactDirsByPath.values()) {
-    for (const artifactPath of listSessionTranscriptArtifactFiles(artifactDir)) {
-      const normalizedArtifactPath = normalizeRealComparablePath(artifactPath);
-      if (scannedArtifactPaths.has(normalizedArtifactPath)) {
-        continue;
-      }
-      scannedArtifactPaths.add(normalizedArtifactPath);
-      if (activeEntriesByPath.has(normalizedArtifactPath)) {
-        continue;
-      }
-      const artifactOwner = activeEntryOwnersByPath.get(normalizedArtifactPath);
-      if (artifactOwner) {
-        continue;
-      }
-      const primarySessionId = parseUsageCountedSessionIdFromFileName(path.basename(artifactPath));
-      const primaryOwner =
-        primarySessionId && isSessionArchiveArtifactName(path.basename(artifactPath))
-          ? activeEntryOwnersByPath.get(
-              normalizeRealComparablePath(
-                path.join(path.dirname(artifactPath), `${primarySessionId}.jsonl`),
-              ),
-            )
-          : undefined;
-      if (primaryOwner && primaryOwner !== normalizedAgentId) {
-        continue;
-      }
-      if (!primaryOwner && !includeUnownedArtifacts) {
-        continue;
-      }
-      const entry = toArtifactCorpusEntry(normalizedAgentId, artifactPath, activeEntriesByPath);
-      if (entry) {
-        corpusEntries.push(entry);
-      }
-    }
-  }
-  return corpusEntries;
-}
-
-/**
- * Lists transcript corpus entries for QMD/memory indexing.
- *
- * Active sessions come from the session accessor seam; retained reset/delete
- * transcript artifacts remain explicit file artifacts until core owns archive
- * artifact enumeration.
- */
-export async function listSessionTranscriptCorpusEntriesForAgent(
-  agentId: string,
-): Promise<SessionTranscriptCorpusEntry[]> {
-  return listSessionTranscriptCorpusEntriesForAgentSync(agentId);
 }
 
 export function loadSessionTranscriptClassificationForSessionsDir(
